separate message functions of the websocket class
